replace index file
