+++ rank/pages/index.tsx
@@ -1,103 +1,34 @@
 import { GetServerSideProps, NextPage } from 'next'
-import Head from 'next/head'
-import absoluteUrl from 'next-absolute-url'
-import { ApiResponse as EvalApiResponse } from './api/eval'
-import Result from '../components/Result'
+import { SearchTemplate, getTemplates } from '../services/search-templates'
+import { removeEmpty } from '../utils'
+import SearchTemplateTests from '../components/SearchTemplateTests'
 
 type Props = {
-  data: EvalApiResponse
-  search: {
-    queryId?: string
-  }
+  searchTemplates: SearchTemplate[]
+}
+
+function get(str: string) {
+  const [env, namespace, index] = str.split('/')
+  return { env, namespace, index }
 }
 
 export const getServerSideProps: GetServerSideProps<Props> = async ({
-  query: qs,
-  req,
+  query,
 }) => {
-  const queryId = qs.queryId ? qs.queryId.toString() : undefined
-  const { origin } = absoluteUrl(req)
-  const resp = await fetch(`${origin}/api/eval?env=prod`)
-  const data: EvalApiResponse = await resp.json()
+  const tests = (
+    Array.isArray(query.test) ? query.test : query.test ? [query.test] : []
+  ).map((test) => get(test))
 
+  const searchTemplates = await getTemplates({ prod: true })
   return {
-    props: {
-      data,
-      search: JSON.parse(
-        JSON.stringify({
-          queryId,
-        })
-      ),
-    },
+    props: removeEmpty({
+      searchTemplates: searchTemplates,
+    }),
   }
 }
 
-type ResultComponentProps = {
-  result: EvalApiResponse['results'][number]
+export const Index2: NextPage<Props> = ({ searchTemplates }) => {
+  return <SearchTemplateTests searchTemplates={searchTemplates} />
 }
 
-const ResultsComponent = ({ result }: ResultComponentProps) => {
-  const { namespace } = result
-  return (
-    <div className="py-4 font-mono">
-      <h2 className="text-2xl font-bold">
-        <Result pass={result.pass} /> {result.label} in {result.namespace}
-      </h2>
-
-      <h3 className="font-bold">Queries:</h3>
-      <ul>
-        {Object.entries(result.results).map(
-          ([key, { query, result, description }]) => {
-            const encodedQuery = encodeURIComponent(query)
-            const searchURL = `https://wellcomecollection.org/${namespace}?query=${encodedQuery}`
-            return (
-              <li key={key} className="py-1">
-                <Result pass={result.pass} /> <a href={searchURL}>{query}</a>{' '}
-                <p className="text-gray-500 text-sm">{description || null}</p>
-              </li>
-            )
-          }
-        )}
-      </ul>
-    </div>
-  )
-}
-
-const Index: NextPage<Props> = ({ data: { results } }: Props) => {
-  return (
-    <>
-      <Head>
-        <title>Relevancy ranking evaluation | Wellcome Collection</title>
-      </Head>
-
-      <h1 className="text-4xl font-bold">Rank eval</h1>
-
-      <p className="py-2">
-        When someone runs a search on{' '}
-        <a href="https://wellcomecollection.org/works">
-          wellcomecollection.org
-        </a>
-        , we transform their search terms into some structured json. That json
-        forms the <i>query</i> which is run against our data in elasticsearch.
-        <br />
-        We update the structure of our queries periodically to improve the
-        relevance of our search results.
-        <br />
-        Every time we update a query, we test it against a set of known search
-        terms, making sure that we{"'"}re always showing people the right stuff.
-        <br />
-        You can{' '}
-        <a href="https://api.wellcomecollection.org/catalogue/v2/search-templates.json">
-          see the current candidate search queries here
-        </a>
-        .
-      </p>
-
-      {results.map((result) => (
-        <ResultsComponent result={result} key={result.label} />
-      ))}
-    </>
-  )
-}
-
-export default Index
+export default Index2
